requirements-worker.txt
# Worker & Tooling Dependencies
torch
faster-whisper
opencc-python-reimplemented
yt-dlp
google-generativeai
pydub
WeasyPrint
